% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_brain.sagittal.R
\docType{data}
\name{brain.sagittal}
\alias{brain.sagittal}
\title{brain.sagittal}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 8 rows and 2 columns.
}
\usage{
data(brain.sagittal)
}
\description{
brain.sagittal higher mental functions, motor function areas, motor function area, sensory area, sensory association area, visual area, Oifactory area, emotional area.
}
\examples{
data(brain.sagittal)
}
\keyword{datasets}
