% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mandible.R
\docType{data}
\name{mandible}
\alias{mandible}
\title{mandible}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 5 rows and 2 columns.
}
\usage{
data(mandible)
}
\description{
mandible outlier data include head, neck, ramus, coronoid and body.
}
\examples{
data(mandible)
}
\keyword{datasets}
